Added functionality to GenBinaryDataFile
The main method of GenBinaryDataFile now takes a third parameter which allows us to specify what type of file we want to create. These are the options I implemented so far:
"random" - creates a randomly populated file (this is what the method already did)
"sorted" - creates a sorted file, with the first value being 0.0 and then incrementing by 0.0001 for each following value.
"reverseSorted" - creates a reverse sorted file (so completely unsorted technically) with the first value being Double.MAX_VALUE and decrementing by 0.0001 for each following value.

I also updated all references to GenBinaryDataFile in the tests you wrote so that they take the parameter "random".